added openmp checker

+++ src/pois_reg_log1p.cpp
@@ -130,7 +130,7 @@ arma::mat regress_cols_of_Y_on_X_log1p_pois_exact(
   if (common_size_factor) {
 
     #pragma omp parallel for shared(B)
-    for (int j = 0; j < B.n_cols; j++) {
+    for (int j = 0; j < static_cast<int>(B.n_cols); j++) {
 
       arma::vec s_j(Y[j].n_elem);
       s_j.fill(s[j]);
@@ -152,7 +152,7 @@ arma::mat regress_cols_of_Y_on_X_log1p_pois_exact(
   } else {
 
     #pragma omp parallel for shared(B)
-    for (int j = 0; j < B.n_cols; j++) {
+    for (int j = 0; j < static_cast<int>(B.n_cols); j++) {
 
       B.col(j) = solve_pois_reg_log1p (
         X,
